+++ src/viewer/Chart.footer.js
@@ -1,3 +1,0 @@
-﻿return {"ChartViewer" : ChartViewer, "Plot" : Plot};
-}); // end of the factory function
-}(window));