+++ src/viewer/chartViewer.d.ts
@@ -1,4 +1,4 @@
-declare module ChartViewer {
+declare module InteractiveDataDisplay {
     export interface ViewState {
         [name: string]: any;
     }
@@ -25,7 +25,11 @@ declare module ChartViewer {
     /**ChartViewer.show() gets a map of pairs (plot identifier, plot definition) as a second argument.
     Plot definition is a collection of key- value pairs specifying properties of a plot, such as line stroke or marker shape.
     Each plot definition has at least one property which is type; it determines a rendering method and must be known to the ChartViewer.*/
-    function show(domElement: HTMLElement, plots: ChartInfo, viewState?: ViewState): ChartViewer.ViewerControl;
+    function show(domElement: HTMLElement, plots: ChartInfo, viewState?: ViewState): InteractiveDataDisplay.ViewerControl;
+
+    export var SizePalette: any;
+    export var ColorPalette: any;
+    export var Utils: any;
 }
 
 declare module Plot {
@@ -137,19 +141,19 @@ declare module Plot {
     };
     /**line draws a grid function y[i] = y(x[i]) where y[i] is either a scalar value or a random variable distribution.
     In former case, a single polyline is drawn; in the latter case, a median polyline along with filled bands for percentiles of the distribution is rendered.*/
-    function line(element: LineDefinition): ChartViewer.PlotInfo;
+    function line(element: LineDefinition): InteractiveDataDisplay.PlotInfo;
     /**The plot draws a colored band between two scalar grid functions.
     The space between lines y1[i](x[i]) and y2[i](x[i]) is filled with a solid color; the boundaries are not stroked.*/
-    function band(element: BandDefinition): ChartViewer.PlotInfo;
+    function band(element: BandDefinition): InteractiveDataDisplay.PlotInfo;
     /**The plot draws a colored boxplot.*/
-    function boxplot(element: BoxPlotDefinition): ChartViewer.PlotInfo;
+    function boxplot(element: BoxPlotDefinition): InteractiveDataDisplay.PlotInfo;
     /**Displays data as a collection of points, each having the value of one variable determining 
     the position on the horizontal axis and the value of the other variable determining the position on the vertical axis. 
     Also variables can be bound to marker size and color. 
     Dependent variable, size-bound variable or color-bound variable 
     can be real or uncertain; the latter is represented as a set of quantiles.*/
-    function markers(element: MarkersDefinition): ChartViewer.PlotInfo;
+    function markers(element: MarkersDefinition): InteractiveDataDisplay.PlotInfo;
     /**Heatmap plot renders values defined on a rectangular grid using color palette*/
-    function heatmap(element: HeatmapDefinition): ChartViewer.PlotInfo;
+    function heatmap(element: HeatmapDefinition): InteractiveDataDisplay.PlotInfo;
 }