+++ test/manual/config.js
@@ -2,13 +2,11 @@ requirejs.config({
     baseUrl: ".",
     paths: {
         "jquery": "../../ext/jquery/dist/jquery",
-        "idd": "../../dist/idd.min",
-        "idd-css": "../../dist/idd",
         "css": "../../ext/require-css/css",
-        "chartViewer": "../../dist/chartViewer",
-        "chartViewer-css": "../../dist/chartViewer",
+        "idd": "../../dist/idd",
+        "idd-css": "../../dist/idd",
         "jquery-ui": "../../ext/jqueryui/jquery-ui",
-        "rx": "../../ext/rxjs/dist/rx.all"
+        "rx": "../../ext/rxjs/dist/rx.all",
     }
 });