+++ RELEASE_NOTES.md
@@ -1,12 +1,12 @@
-##### 1.3.1 (Unpublished)
 
+
+##### 1.3.1 (Unpublished)
 New features:
  - IDD release includes idd.ko.js which is a superset for idd.js with basic support for Knockout. See coming examples in the [Wiki](https://github.com/predictionmachines/InteractiveDataDisplay/wiki).
 
 Changes:
  - Box-and-whisker markers are now represented by a single shape `"boxwhisker"` and the actual marker type (either box-and-whisker, box or just whisker) depends on the drawn data. See the [samples](https://github.com/predictionmachines/InteractiveDataDisplay/blob/master/test/manual/Boxwhisker%20plot.html).
-
-
+ - **remove files chartViewer.js (.css, .d.ts). Concat idd.js and chartViewer.js. Add idd.umd.js. Change namespace ChartViewer on InteractiveDataDisplay.**
 ## 1.3.0 (May 27, 2016)
 
 New features: